i156 Tweak the scope concept for EntityManagerFactory retrieval so that defaults are easier to the user
Now when no scope is provided, we will try to retrieve an entity manager
factory by any means that make sense for the current context (CDI/non-CDI):
we won't be limited by the default scope. This means in particular that
users who created a single SessionFactory using the Hibernate APIs will
still be able to use this factory as a default (before this change, only
factories created through the JPA APIs would be be detected by default).

The main change is that there is now only one registry per context
(CDI/non-CDI), allowing this registry to use more methods to get the
default entity manager factory.

Incidentally, this means that it's no longer possible to use non-CDI
scopes in a CDI environment. It doesn't seem too bad, but if necessary
we can restore the old behavior easily anyway (by making the CDI
registry delegate to the non-CDI one).
