+++ core/src/main/java/org/hibernate/search/jsr352/context/jpa/EntityManagerFactoryRegistry.java
@@ -9,21 +9,34 @@ package org.hibernate.search.jsr352.context.jpa;
 import javax.persistence.EntityManagerFactory;
 
 /**
- * An abstract contract allowing to retrieve an entity manager factory
- * without any assumption about the underlying dependency injection mechanism.
+ * An abstract contract allowing to retrieve an entity manager factory.
  *
  * @author Yoann Rodiere
  */
 public interface EntityManagerFactoryRegistry {
 
 	/**
-	 * @return The default {@link EntityManagerFactory}.
+	 * @return The default {@link EntityManagerFactory}, if there is one.
 	 */
 	EntityManagerFactory getDefault();
 
 	/**
+	 * Retrieve a factory using the default (implementation-dependent) scope.
+	 *
+	 * @param reference The reference allowing to identify the factory uniquely. Must be non-null and non-empty.
 	 * @return The {@link EntityManagerFactory} for the given reference string.
 	 */
 	EntityManagerFactory get(String reference);
+
+	/**
+	 * Retrieve a factory using the given scope.
+	 *
+	 * @param scope The scope of the reference; accepted scopes are implementation-dependent. Must be non-null and non-empty.
+	 * For instance an implementation could accept the scope 'persistence-unit-name', meaning
+	 * that the reference will be interpreted as a persistence unit name.
+	 * @param reference The reference allowing to identify the factory uniquely. Must be non-null and non-empty.
+	 * @return The {@link EntityManagerFactory} for the given reference string.
+	 */
+	EntityManagerFactory get(String scope, String reference);
 
 }
